Modificaciones Users

+++ backendProject/src/main/java/com/bubblebear/backendProject/Repository/UsersRepository.java
@@ -6,10 +6,16 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.bubblebear.backendProject.entity.Users;
 
-public interface UsersRepository extends CrudRepository<Users, Long> {
+public interface UsersRepository extends CrudRepository<Users, Integer> {
 		
 		List<Users> findByFullName(String fullName);
 		
 		Users findById(long id);
-	
+
+		//Users creatUsers(Users usersDto);
+
+		//List<Users> getAllUsers();
+
+		//Users updateUsers(Users usersDto, long id);
+
 }
